fix: fix box.sql.execute, schema, infinite retries
In older tarantool versions querying with sql was supposed to be done
with box.sql.execute, however it was changed to box.execute in newer
ones. This lead to not being able to create a valid snapshot on newer
version of Tarantool. SQL compatibility also changed over the years
so the types were also changed. Retries were handled very dirty, so
I also added a threshold for tpcc_load not to enter an infinite loop
that is very unpleasant to debug